LayoutTests/fast/js/sputnik/Conformance/12_Statement/12.8_The_break_Statement/S12.8_A5_T2-expected.txt

CONSOLE MESSAGE: line 82: SyntaxError: Parse error
S12.8_A5_T2

PASS Expected parsing failure

TEST COMPLETE
